Reordered the gameplay
I realized that all of the gameplay needed to be contained inside a function, and each round needed to be a function inside that, so that it could execute multiple times. I also moved the "toLowerCase" method so that it only executes when the player enters a value, otherwise it tries to operate on "null" which renders an error. I now need to figure out how to make the entire thing loop without having to reload the page, which isn't required, but I'd like to do it anyway.