Combined similar selection conditions
Realized that I didn't need to have three separate conditions for if computerSelection && playerSelection = rock, paper, or scissors, I only needed one condition that checks if those two variables are the same. This simplifies the code from nine separate condition checks to seven, which isn't a ton, but every bit of simplification counts.