Beginning stage of adding ui
I added a functioning UI to the rock paper scissors game, so it doesn't have to just be played in the console window. I created buttons for the player selection, and DOM manipulations that display the result and update the score as they go along. Once either the computer or the player's score reaches 5, it displays a message stating who won the game. Now trying to work out how to get the game to restart/reload once this happens, without the user having to reload the whole page themselves and preventing them from continuing to play once that score is reached.