Fixed spacing issue
I was having an issue with spacing of the paragraphs when they were empty, before the JS added text. I needed them to take up the same amount of space empty so that the other elements don't shift. Originall fixed it by having a period in each paragraph set to the same color as the background, then the JS would change the color when the text was inputed, which was a convoluted solution. Realized I just needed to set the paragraphs' height to 1rem in the CSS so that they would take up the same amount of room empty or full. Solved! Also changed the colors in this version so everything is just shades of blue. Cleaner.